Added animation to expand button and changed its colour.
TODO: Figure out using an icon for the arrow. I tried, couldn't make it work easily.
TODO: If possible, implement closing animation for sidebar. Current state of knowledge says that's impossible with the current setup, as the sidebar must be hidden one way or another to prevent it showing through the cursor detector at all times.