Code clean-up
Much cleaner and more easily maintainable code for the export functionality. Now asks for the location/filename when exporting a pdf. Replaced static div with .renderStaticMarkup when exporting to pdf.
TODO: need a way to get the page-divs only -> how tf do I do that in react? works with plain js...